// src/app/core/models/my-contacts.model.ts
export class MyContacts {
      uuid: string;
      id: string;
      first_name: string;
      phone_dial_code: string;
      phone_number: string;
      email: string;
      advisor_id_number: number;
      advisor_role: string;
}



export class CustomerInforAdvisor {
      id: string;
      first_name: string;
      phone_dial_code: string;
      phone_number: string;
      email: string;
      address_line1: string;
      address_line2: string;
      address_line3: string;
      postal_code: string;
      state_code: string;
      country_code: string;
      customer_id : number;
}